- move mock data and utils - use functions instead of constants
