+++ frontend/app/component/form/data_entry/check_and_save/CheckAndSaveForm.test.tsx
@@ -14,51 +14,55 @@ import { renderReturningRouter, screen, server, spyOnHandler, within } from "@ki
 import { DataEntryProvider } from "../state/DataEntryProvider";
 import { DataEntryState } from "../state/types";
 import {
-  defaultFormSection,
-  emptyDataEntryRequest,
   errorWarningMocks,
+  getDefaultFormSection,
+  getEmptyDataEntryRequest,
   overrideServerGetDataEntryResponse,
-} from "../test.util";
+} from "../test-data";
 import { CheckAndSaveForm } from "./CheckAndSaveForm";
 
-const defaultValues = emptyDataEntryRequest.data;
+function getDefaultValues() {
+  return getEmptyDataEntryRequest().data;
+}
 
-const defaultDataEntryState: DataEntryState = {
-  election: electionMockData,
-  pollingStationId: 1,
-  error: null,
-  pollingStationResults: null,
-  entryNumber: 1,
-  formState: {
-    current: "save",
-    furthest: "save",
-    sections: {
-      recounted: {
-        id: "recounted",
-        index: 1,
-        ...defaultFormSection,
-      },
-      voters_votes_counts: {
-        id: "voters_votes_counts",
-        index: 2,
-        ...defaultFormSection,
-      },
-      differences_counts: {
-        id: "differences_counts",
-        index: 3,
-        ...defaultFormSection,
-      },
-      save: {
-        id: "save",
-        index: 4,
-        ...defaultFormSection,
+function getDefaultDataEntryState(): DataEntryState {
+  return {
+    election: electionMockData,
+    pollingStationId: 1,
+    error: null,
+    pollingStationResults: null,
+    entryNumber: 1,
+    formState: {
+      current: "save",
+      furthest: "save",
+      sections: {
+        recounted: {
+          id: "recounted",
+          index: 1,
+          ...getDefaultFormSection(),
+        },
+        voters_votes_counts: {
+          id: "voters_votes_counts",
+          index: 2,
+          ...getDefaultFormSection(),
+        },
+        differences_counts: {
+          id: "differences_counts",
+          index: 3,
+          ...getDefaultFormSection(),
+        },
+        save: {
+          id: "save",
+          index: 4,
+          ...getDefaultFormSection(),
+        },
       },
     },
-  },
-  targetFormSectionId: "recounted",
-  status: "idle",
-  cache: null,
-};
+    targetFormSectionId: "recounted",
+    status: "idle",
+    cache: null,
+  };
+}
 
 function renderForm() {
   return renderReturningRouter(
@@ -113,8 +117,8 @@ describe("Test CheckAndSaveForm", () => {
 
   test("Data entry does not show finalise button with errors", async () => {
     overrideServerGetDataEntryResponse({
-      formState: defaultDataEntryState.formState,
-      pollingStationResults: defaultValues,
+      formState: getDefaultDataEntryState().formState,
+      pollingStationResults: getDefaultValues(),
       validationResults: {
         errors: [
           {
@@ -141,8 +145,8 @@ describe("Test CheckAndSaveForm", () => {
 
   test("Data entry does not show finalise button with unaccepted warnings", async () => {
     overrideServerGetDataEntryResponse({
-      formState: defaultDataEntryState.formState,
-      pollingStationResults: defaultValues,
+      formState: getDefaultDataEntryState().formState,
+      pollingStationResults: getDefaultValues(),
       validationResults: {
         errors: [],
         warnings: [
@@ -163,12 +167,12 @@ describe("Test CheckAndSaveForm", () => {
   });
 
   test("Data entry shows finalise button with accepted warnings", async () => {
-    const dataEntryState = structuredClone(defaultDataEntryState);
+    const dataEntryState = getDefaultDataEntryState();
     dataEntryState.formState.sections.voters_votes_counts.acceptWarnings = true;
 
     overrideServerGetDataEntryResponse({
       formState: dataEntryState.formState,
-      pollingStationResults: defaultValues,
+      pollingStationResults: getDefaultValues(),
       validationResults: {
         errors: [],
         warnings: [
@@ -191,10 +195,10 @@ describe("Test CheckAndSaveForm summary", () => {
     server.use(ElectionRequestHandler, PollingStationDataEntryGetHandler, PollingStationDataEntrySaveHandler);
   });
   test("Blocking", async () => {
-    const values = structuredClone(defaultValues);
+    const values = getDefaultValues();
 
     overrideServerGetDataEntryResponse({
-      formState: defaultDataEntryState.formState,
+      formState: getDefaultDataEntryState().formState,
       pollingStationResults: values,
       validationResults: {
         errors: [errorWarningMocks.F201],
@@ -221,11 +225,11 @@ describe("Test CheckAndSaveForm summary", () => {
   });
 
   test("Accepted with warnings", async () => {
-    const dataEntryState = structuredClone(defaultDataEntryState);
+    const dataEntryState = getDefaultDataEntryState();
     dataEntryState.formState.sections.differences_counts.acceptWarnings = true;
     overrideServerGetDataEntryResponse({
       formState: dataEntryState.formState,
-      pollingStationResults: defaultValues,
+      pollingStationResults: getDefaultValues(),
       validationResults: {
         errors: [],
         warnings: [errorWarningMocks.W301],
@@ -246,8 +250,8 @@ describe("Test CheckAndSaveForm summary", () => {
 
   test("Unaccepted warnings", async () => {
     overrideServerGetDataEntryResponse({
-      formState: defaultDataEntryState.formState,
-      pollingStationResults: defaultValues,
+      formState: getDefaultDataEntryState().formState,
+      pollingStationResults: getDefaultValues(),
       validationResults: {
         errors: [],
         warnings: [errorWarningMocks.W301],
