+++ frontend/app/component/form/data_entry/state/reducer.test.ts
@@ -4,18 +4,24 @@ import { mockElection } from "app/component/election/status/mockData";
 
 import { ApiResponseStatus } from "@kiesraad/api";
 
-import { getInitialValues } from "./dataEntryUtils";
-import dataEntryReducer, { getInitialState } from "./reducer";
+import { getInitialValues as _getInitialValues } from "./dataEntryUtils";
+import dataEntryReducer, { getInitialState as _getInitialState } from "./reducer";
 import { DataEntryAction, DataEntryState } from "./types";
 
-const initialState: DataEntryState = getInitialState(mockElection, 1, 1);
-const initialValues = getInitialValues(mockElection);
+function getInitialState(): DataEntryState {
+  return _getInitialState(mockElection, 1, 1);
+}
+
+function getInitialValues() {
+  return _getInitialValues(mockElection);
+}
+
 test("should handle DATA_ENTRY_LOADED with client_state", () => {
   const action: DataEntryAction = {
     type: "DATA_ENTRY_LOADED",
     dataEntry: {
       client_state: null,
-      data: initialValues,
+      data: getInitialValues(),
       progress: 0,
       validation_results: {
         errors: [],
@@ -24,7 +30,7 @@ test("should handle DATA_ENTRY_LOADED with client_state", () => {
       updated_at: "",
     },
   };
-  const state = dataEntryReducer(initialState, action);
+  const state = dataEntryReducer(getInitialState(), action);
   expect(state.formState).toBeDefined();
   expect(state.targetFormSectionId).toBeDefined();
   expect(state.pollingStationResults).toEqual(action.dataEntry.data);
@@ -36,9 +42,9 @@ test("should handle DATA_ENTRY_NOT_FOUND", () => {
     type: "DATA_ENTRY_NOT_FOUND",
   };
 
-  const state = dataEntryReducer(initialState, action);
+  const state = dataEntryReducer(getInitialState(), action);
   expect(state.error).toBeNull();
-  expect(state.pollingStationResults).toStrictEqual(initialValues);
+  expect(state.pollingStationResults).toStrictEqual(getInitialValues());
 });
 
 test("should handle DATA_ENTRY_LOAD_FAILED", () => {
@@ -53,7 +59,7 @@ test("should handle DATA_ENTRY_LOAD_FAILED", () => {
     },
   };
 
-  const state = dataEntryReducer(initialState, action);
+  const state = dataEntryReducer(getInitialState(), action);
   expect(state.error).toBeDefined();
   expect(state.error).toEqual(action.error);
 });
@@ -64,7 +70,7 @@ test("should handle SET_STATUS", () => {
     status: "saving",
   };
 
-  const state = dataEntryReducer(initialState, action);
+  const state = dataEntryReducer(getInitialState(), action);
   expect(state.status).toBeDefined();
   expect(state.status).toEqual(action.status);
 });
@@ -80,13 +86,13 @@ test("should handle SET_CACHE", () => {
     },
   };
 
-  const state = dataEntryReducer(initialState, action);
+  const state = dataEntryReducer(getInitialState(), action);
   expect(state.cache).toBeDefined();
   expect(state.cache).toEqual(action.cache);
 });
 
 test("should handle UPDATE_FORM_SECTION", () => {
-  const oldState = structuredClone(initialState);
+  const oldState = getInitialState();
   oldState.formState.current = "voters_votes_counts";
 
   const action: DataEntryAction = {
@@ -113,7 +119,7 @@ test("should handle FORM_SAVE_FAILED", () => {
     },
   };
 
-  const state = dataEntryReducer(initialState, action);
+  const state = dataEntryReducer(getInitialState(), action);
   expect(state.error).toBeDefined();
   expect(state.error).toEqual(action.error);
 });
@@ -121,7 +127,7 @@ test("should handle FORM_SAVE_FAILED", () => {
 test("should handle FORM_SAVED", () => {
   const action: DataEntryAction = {
     type: "FORM_SAVED",
-    data: initialValues,
+    data: getInitialValues(),
     validationResults: {
       errors: [],
       warnings: [],
@@ -130,7 +136,7 @@ test("should handle FORM_SAVED", () => {
     continueToNextSection: true,
   };
 
-  const state = dataEntryReducer(initialState, action);
+  const state = dataEntryReducer(getInitialState(), action);
   expect(state.error).toBeNull();
   expect(state.pollingStationResults).toEqual(action.data);
   expect(state.targetFormSectionId).toBeDefined();
@@ -138,7 +144,7 @@ test("should handle FORM_SAVED", () => {
 });
 
 test("should handle RESET_TARGET_FORM_SECTION", () => {
-  const oldState = structuredClone(initialState);
+  const oldState = getInitialState();
   oldState.targetFormSectionId = "voters_votes_counts";
 
   const action: DataEntryAction = {
@@ -158,7 +164,7 @@ test("should handle REGISTER_CURRENT_FORM", () => {
     },
   };
 
-  const state = dataEntryReducer(initialState, action);
+  const state = dataEntryReducer(getInitialState(), action);
   expect(state.formState.current).toBeDefined();
   expect(state.formState.current).toEqual(action.form.id);
   expect(state.formState.sections.recounted.isSubmitted).toBeDefined();
