+++ frontend/app/component/form/data_entry/test-data/index.ts
@@ -1,0 +1,2 @@
+export * from "./mock-data";
+export * from "./test.utils";
